security issue corrected

+++ src/main/java/lk/learners/karunadasa/transactions/finance/entity/Invoice.java
@@ -1,9 +1,8 @@
 package lk.learners.karunadasa.transactions.finance.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lk.learners.karunadasa.Security.entity.User;
-import lk.learners.karunadasa.resourses.company.entity.Student;
-import lk.learners.karunadasa.transactions.finance.entity.DiscountRatio;
+import lk.learners.karunadasa.general.security.entity.User;
+import lk.learners.karunadasa.resourses.company.human.entity.Student;
 import lk.learners.karunadasa.transactions.finance.entity.Enum.InvoicePrintOrNot;
 import lk.learners.karunadasa.transactions.finance.entity.Enum.PaymentMethod;
 import lombok.*;
@@ -25,10 +24,10 @@ public class Invoice {
     @Id
     @Column(name = "id", nullable = false, unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "number", nullable = false, unique = true)
-    private Integer number;
+    private Long number;
 
     @Column(name = "payment_method", nullable = false, length = 10)
     @Enumerated(EnumType.STRING)
@@ -56,7 +55,7 @@ public class Invoice {
 
 
     @Column(name = "card_number")
-    private Integer cardNumber;
+    private Long cardNumber;
 
 
     @Column(name = "remarks", length = 150)
